+++ tests/test_backend_opencv.py
@@ -22,6 +22,7 @@ import pytest
 
 from scenedetect import ContentDetector, SceneManager
 from scenedetect.backends.opencv import VideoCaptureAdapter, VideoStreamCv2
+from scenedetect.video_stream import SeekError
 
 GROUND_TRUTH_CAPTURE_ADAPTER_TEST = [1, 90, 210]
 GROUND_TRUTH_CAPTURE_ADAPTER_CALLBACK_TEST = [180, 394]
@@ -35,7 +36,9 @@ def test_open_image_sequence(test_image_sequence: str):
     assert sequence.duration is not None
     assert sequence.duration.frame_num == 30
     assert sequence.read() is not False
-    sequence.seek(100)
+    with pytest.raises(SeekError):
+        sequence.seek(100)
+    sequence.seek(sequence.duration.frame_num - 1)
     assert sequence.position == 29